move eos to progress task
